pyproject.toml
[project]
name = "multimodal-registration"
version = "0.1.0"
description = "Multimodal 3-D image registration (DCT + PCT) with optional GPU acceleration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "h5py>=3.8",
    "scikit-image>=0.21",
    "matplotlib>=3.7",
]

[project.optional-dependencies]
# VTK file support for DVC results
vtk = ["pyvista>=0.43"]
# Jupyter interactive registration UI
jupyter = ["ipywidgets>=8.0", "ipympl>=0.9"]
# Crystallographic orientation maths (IPF update after DVC)
crystal = ["orix>=0.12"]
# NVIDIA GPU acceleration — pick the wheel matching your CUDA version:
#   pip install "multimodal-registration[cuda12]"
cuda11 = ["cupy-cuda11x>=12.0", "cucim-cu11>=23.0"]
cuda12 = ["cupy-cuda12x>=12.0", "cucim-cu12>=23.0"]

[project.scripts]
multimodal-registration = "multimodal_registration:main"

[dependency-groups]
docs = ["mkdocs>=1.6", "mkdocstrings[python]>=0.24", "pymdown-extensions>=10.0"]

[build-system]
requires = ["uv_build>=0.8.9,<0.9.0"]
build-backend = "uv_build"
